fotothis/requirements.txt
functions-framework==3.*
flask==3.0.2
google-cloud-storage==2.16.0
PyYAML==6.0.1
